[Bronze II] Title: 숫자, Time: 28 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ C++17/백준/Bronze/10093. 숫자/숫자.cc
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    unsigned int a, b, c;
+    long long a, b, c;
     cin >> a>> b;
     if (a==b){
         cout << 0;
@@ -15,7 +15,7 @@ int main(){
     }
     cout << b-a-1<<endl;
 
-    for (unsigned int i=a+1;i<b;i++){
+    for (long long i=a+1;i<b;i++){
         cout << i<<" ";
     }
